Refactor to separate out different routes
